Fixing failing mac_brew tests.  Moving to pytest.
